requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
ollama>=0.1.6
pdfplumber>=0.11.0
python-multipart>=0.0.9
requests>=2.31.0
pydantic>=2.6.3
groq>=0.4.0
psycopg2-binary
python-dotenv
